Handle new "macro-value" span element in entries
Regular diary entries, and the goals entries, had a span element added
to support toggling views of %age based macro goals and raw nutrition values.

Adds support for getting the actual values from the table of a diary page.
Ignores the %age values but these could be implemented in future.

Also adds some span elements to the test page to ensure the changes work as
expected.